requirements.txt
ftfy
omegaconf
regex
scikit-learn
submitit
termcolor
torch
torchmetrics
torchvision
transformers
Pillow
numpy
accelerate
tqdm
